refactor: combine process threshold function into the prober class
