+++ src/components/notification/__tests__/process-server-status.test.ts
@@ -22,17 +22,9 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
-import { expect } from 'chai'
+import { expect } from '@oclif/test'
 import { interpret } from 'xstate'
-
-import { Probe } from '../../../interfaces/probe'
-import { ServerAlertState } from '../../../interfaces/probe-status'
-import { ValidatedResponse } from '../../../plugins/validate-response'
-import {
-  processThresholds,
-  serverAlertStateInterpreters,
-  serverAlertStateMachine,
-} from '../process-server-status'
+import { serverAlertStateMachine } from '../process-server-status'
 
 describe('serverAlertStateMachine', () => {
   let interpreter: any
@@ -122,162 +114,5 @@ describe('serverAlertStateMachine', () => {
       'SUCCESS',
     ])
     expect(interpreter.state.value).to.equals('DOWN')
-  })
-})
-
-describe('processThresholds', () => {
-  beforeEach(() => {
-    serverAlertStateInterpreters.clear()
-  })
-
-  const probe = {
-    requests: [
-      {
-        method: 'GET',
-        url: 'https://httpbin.org/status/200',
-      },
-      {
-        method: 'POST',
-        url: 'https://httpbin.org/status/201',
-      },
-    ],
-    incidentThreshold: 2,
-    recoveryThreshold: 2,
-  } as Probe
-
-  const successResponse = [
-    {
-      alert: { query: 'response.time > 1000' },
-      isAlertTriggered: false,
-    },
-  ] as ValidatedResponse[]
-  const failureResponse = [
-    {
-      alert: { query: 'response.time > 1000' },
-      isAlertTriggered: true,
-    },
-  ] as ValidatedResponse[]
-
-  it('should attach state calculation to each request', () => {
-    // failure happened first for request with index 0
-    processThresholds({
-      probe,
-      requestIndex: 0,
-      validatedResponse: failureResponse,
-    })
-
-    // processing request with index 1
-    const result1 = processThresholds({
-      probe,
-      requestIndex: 1,
-      validatedResponse: failureResponse,
-    })
-
-    // failure happened only once for request with index 1, it does not reach threshold yet
-    expect(result1[0].state).to.equals('UP')
-
-    const result2 = processThresholds({
-      probe,
-      requestIndex: 1,
-      validatedResponse: failureResponse,
-    })
-
-    // second time failure happened for request with index 1, it reaches threshold
-    expect(result2[0].state).to.equals('DOWN')
-  })
-
-  it('should compute shouldSendNotification based on threshold and previous state', () => {
-    let result!: ServerAlertState[]
-
-    // trigger down state
-    for (let i = 0; i < probe.incidentThreshold; i++) {
-      result = processThresholds({
-        probe,
-        requestIndex: 1,
-        validatedResponse: failureResponse,
-      })
-    }
-
-    expect(result[0].state).to.equals('DOWN')
-
-    // initial state is UP and now it is changed to DOWN, should send notification
-    expect(result[0].shouldSendNotification).to.equals(true)
-
-    // send success response once
-    result = processThresholds({
-      probe,
-      requestIndex: 1,
-      validatedResponse: successResponse,
-    })
-
-    // send failure response again as much threshold
-    for (let i = 0; i < probe.incidentThreshold; i++) {
-      result = processThresholds({
-        probe,
-        requestIndex: 1,
-        validatedResponse: failureResponse,
-      })
-    }
-
-    expect(result[0].state).to.equals('DOWN')
-
-    // should not send notification again since state does not change
-    expect(result[0].shouldSendNotification).to.equals(false)
-
-    // send success response as much threshold
-    for (let i = 0; i < probe.recoveryThreshold; i++) {
-      result = processThresholds({
-        probe,
-        requestIndex: 1,
-        validatedResponse: successResponse,
-      })
-    }
-
-    expect(result[0].state).to.equals('UP')
-
-    // should send notification since state change from DOWN to UP
-    expect(result[0].shouldSendNotification).to.equals(true)
-
-    // send failure response once
-    result = processThresholds({
-      probe,
-      requestIndex: 1,
-      validatedResponse: failureResponse,
-    })
-
-    // send success response as much threshold
-    for (let i = 0; i < probe.recoveryThreshold; i++) {
-      result = processThresholds({
-        probe,
-        requestIndex: 1,
-        validatedResponse: successResponse,
-      })
-    }
-
-    expect(result[0].state).to.equals('UP')
-
-    // should not send notification again since state does not change
-    expect(result[0].shouldSendNotification).to.equals(false)
-  })
-
-  it('should accept socket probe', () => {
-    // arrange
-    const probe = {
-      id: '_uwBu',
-      socket: {
-        host: 'localhost',
-        port: '5432',
-      },
-    } as unknown as Probe
-
-    // act
-    const result = processThresholds({
-      probe,
-      requestIndex: 1,
-      validatedResponse: failureResponse,
-    })
-
-    // assert
-    expect(result[0].state).eq('UP')
   })
 })
