+++ src/components/probe/prober/index.ts
@@ -23,6 +23,7 @@
  **********************************************************************************/
 
 import type { Notification } from '@hyperjumptech/monika-notification'
+import { interpret } from 'xstate'
 import { checkThresholdsAndSendAlert } from '..'
 import { getContext } from '../../../context'
 import events from '../../../events'
@@ -33,8 +34,12 @@ import { getEventEmitter } from '../../../utils/events'
 import { log } from '../../../utils/pino'
 import { isSymonModeFrom } from '../../config'
 import { RequestLog } from '../../logger'
-import { processThresholds } from '../../notification/process-server-status'
 import responseChecker from '../../../plugins/validate-response/checkers'
+import type { ServerAlertState } from '../../../interfaces/probe-status'
+import {
+  serverAlertStateInterpreters,
+  serverAlertStateMachine,
+} from '../../notification/process-server-status'
 
 export type ProbeResult = {
   isAlertTriggered: boolean
@@ -47,6 +52,11 @@ type RespProsessingParams = {
   probeResults: ProbeResult[]
 }
 
+type ProcessThresholdsParams = {
+  requestIndex: number
+  validatedResponse: ValidatedResponse[]
+}
+
 export interface Prober {
   probe: () => Promise<void>
   generateVerboseStartupMessage: () => string
@@ -54,6 +64,10 @@ export interface Prober {
     response: ProbeRequestResponse,
     additionalAssertions?: ProbeAlert[]
   ) => ValidatedResponse[]
+  processThresholds: ({
+    requestIndex,
+    validatedResponse,
+  }: ProcessThresholdsParams) => ServerAlertState[]
 }
 
 export type ProberMetadata = {
@@ -102,6 +116,71 @@ export class BaseProber implements Prober {
     }))
   }
 
+  processThresholds = ({
+    requestIndex,
+    validatedResponse,
+  }: ProcessThresholdsParams): ServerAlertState[] => {
+    const { requests, incidentThreshold, recoveryThreshold, socket, name } =
+      this.probeConfig
+    const request = requests?.[requestIndex]
+
+    const id = `${this.probeConfig?.id}:${name}:${requestIndex}:${
+      request?.id || ''
+    }-${incidentThreshold}:${recoveryThreshold} ${
+      request?.url || (socket ? `${socket.host}:${socket.port}` : '')
+    }`
+
+    const results: Array<ServerAlertState> = []
+
+    if (!serverAlertStateInterpreters.has(id!)) {
+      const interpreters: Record<string, any> = {}
+
+      for (const alert of validatedResponse.map((r) => r.alert)) {
+        const stateMachine = serverAlertStateMachine.withContext({
+          incidentThreshold,
+          recoveryThreshold,
+          consecutiveFailures: 0,
+          consecutiveSuccesses: 0,
+          isFirstTimeSendEvent: true,
+        })
+
+        interpreters[alert.assertion] = interpret(stateMachine).start()
+      }
+
+      serverAlertStateInterpreters.set(id!, interpreters)
+    }
+
+    // Send event for successes and failures to state interpreter
+    // then get latest state for each alert
+    for (const validation of validatedResponse) {
+      const { alert, isAlertTriggered } = validation
+      const interpreter = serverAlertStateInterpreters.get(id!)![
+        alert.assertion
+      ]
+
+      const prevStateValue = interpreter.state.value
+
+      interpreter.send(isAlertTriggered ? 'FAILURE' : 'SUCCESS')
+
+      const stateValue = interpreter.state.value
+      const stateContext = interpreter.state.context
+
+      results.push({
+        isFirstTime: stateContext.isFirstTimeSendEvent,
+        alertQuery: alert.assertion,
+        state: stateValue as 'UP' | 'DOWN',
+        shouldSendNotification:
+          stateContext.isFirstTimeSendEvent ||
+          (stateValue === 'DOWN' && prevStateValue === 'UP') ||
+          (stateValue === 'UP' && prevStateValue === 'DOWN'),
+      })
+
+      interpreter.send('FIST_TIME_EVENT_SENT')
+    }
+
+    return results
+  }
+
   protected processProbeResults(probeResults: ProbeResult[]): void {
     for (const index of probeResults.keys()) {
       this.responseProcessing({
@@ -126,8 +205,7 @@ export class BaseProber implements Prober {
     const isVerbose = isSymonMode || flags['keep-verbose-logs']
     const validatedResponse = this.validateResponse(probeResult)
     const requestLog = new RequestLog(this.probeConfig, index, 0)
-    const statuses = processThresholds({
-      probe: this.probeConfig,
+    const statuses = this.processThresholds({
       requestIndex: index,
       validatedResponse,
     })
